Further consolodated runnables and widgets into their own files. Improved appearance of QPushButtons.
