+++ README.html
@@ -14,7 +14,7 @@
 	<h2 id="what-am-i">What Am I?</h2>
 	<p>Sermon Prep Database is a program to organize - and store for easy retrieval - your thoughts and research when preparing a sermon.</p>
 
-	<h2>What's New in v.5.0.3.001?</h2>
+	<h2>What's New in v.5.0.3.002?</h2>
 	<ul>
 		<li>Fixed spell-check not happening while typing</li>
 		<li>Fixed misspelled word remaining red after correcting it manually</li>
@@ -49,7 +49,7 @@
 	
 	<h2 id="installation">Installation</h2>
 	<p>Currently, Sermon Prep Database is available for the Microsoft Windows operating
-	system only. Download the current SPD installer (v.5.0.3.001) and run
+	system only. Download the current SPD installer (v.5.0.3.002) and run
 	it on your computer.</p>
 	
 	<h2 id="using-sermon-prep-database">Using Sermon Prep Database</h2>
