+++ compile scripts/SPD Inno Setup Script.iss
@@ -2,7 +2,7 @@
 ; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
 
 #define MyAppName "Sermon Prep Database"
-#define MyAppVersion "v.5.0.3.001"
+#define MyAppVersion "v.5.0.3.002"
 #define MyAppURL "https://sourceforge.net/projects/sermon-prep-database"
 #define MyAppExeName "SermonPrepDatabase.exe"
 #define OutputLocation "C:\Users\pasto\Desktop\output\SermonPrepDatabase"
