+++ gui.py
@@ -1,19 +1,17 @@
-import re
 from os.path import exists
 
-from PyQt6.QtCore import Qt, QSize, QDate, QDateTime, QRunnable, pyqtSignal, QSizeF, QThreadPool
-from PyQt6.QtGui import QIcon, QFont, QStandardItemModel, QStandardItem, QPixmap, QColor, \
-    QCloseEvent, QAction, QUndoStack, QTextCursor, QPainter, QTextDocument, QTextOption, QTextBlockFormat
+from PyQt6.QtCore import Qt, QSize, QDate, QDateTime, pyqtSignal, QThreadPool
+from PyQt6.QtGui import QIcon, QFont, QPixmap, QColor, QCloseEvent, QAction, QUndoStack, QTextCursor, QTextBlockFormat
 from PyQt6.QtWidgets import QWidget, QTabWidget, QGridLayout, QLabel, QLineEdit, \
-    QCheckBox, QDateEdit, QTextEdit, QMainWindow, QVBoxLayout, QHBoxLayout, QPushButton, QTableView, QDialog, \
-    QApplication, QProgressBar, QTabBar
+    QCheckBox, QDateEdit, QTextEdit, QMainWindow, QVBoxLayout, QPushButton, QTabBar
 
 from get_scripture import GetScripture
-from menu_bar import MenuBar
-from toolbar import Toolbar
-from runnables import SpellCheck, LoadDictionary
+from widgets import MenuBar
+from runnables import InitialStartup
+from widgets import Toolbar
+from widgets import CustomTextEdit, ScriptureBox, SermonView
 
-# ToDo: Fix popup buttons
+
 class GUI(QMainWindow):
     main = None
     undo_stack = None
@@ -809,643 +807,3 @@ class GUI(QMainWindow):
         elif event.modifiers() & Qt.KeyboardModifier.ControlModifier and event.key() == Qt.Key.Key_Q:
             self.close()
         event.accept()
-
-
-class InitialStartup(QRunnable):
-    """
-    QDialog class that shows the user the process of starting up the application.
-
-    :param GUI gui: The GUI class.
-    """
-    def __init__(self, gui):
-        super().__init__()
-
-        self.gui = gui
-        self.startup_splash = StartupSplash(gui, 6)
-
-    def run(self):
-        self.startup_splash.update_text.emit('Getting System Info')
-        self.gui.main.get_system_info()
-
-        self.startup_splash.update_text.emit('Getting User Settings')
-        self.gui.main.get_user_settings()
-
-        self.startup_splash.update_text.emit('Loading Dictionaries')
-        self.gui.main.spell_check_thread_pool = QThreadPool()
-        self.gui.main.spell_check_thread_pool.setStackSize(256000000)
-        self.gui.main.load_dictionary_thread_pool = QThreadPool()
-        ld = LoadDictionary(self.gui.main)
-        self.gui.main.load_dictionary_thread_pool.start(ld)
-
-        self.startup_splash.update_text.emit('Getting Indices')
-        self.gui.main.get_ids()
-        self.gui.main.get_date_list()
-        self.gui.main.get_scripture_list()
-        self.gui.main.backup_db()
-
-        self.startup_splash.update_text.emit('Finishing Up')
-        self.gui.create_main_gui.emit()
-        self.startup_splash.end.emit()
-
-        self.gui.changes = False
-
-
-class StartupSplash(QDialog):
-    update_text = pyqtSignal(str)
-    end = pyqtSignal()
-
-    def __init__(self, gui, progress_end):
-        super().__init__()
-        self.gui = gui
-        self.update_text.connect(self.change_text)
-        self.end.connect(lambda: self.done(0))
-
-        self.setWindowFlag(Qt.WindowType.FramelessWindowHint)
-        self.setModal(True)
-        self.setAttribute(Qt.WidgetAttribute.WA_TranslucentBackground)
-        self.setMinimumWidth(300)
-
-        layout = QGridLayout()
-        self.setLayout(layout)
-
-        self.working_label = QLabel()
-        self.working_label.setAutoFillBackground(False)
-        self.working_label.setPixmap(QPixmap('resources/icon.png'))
-        layout.addWidget(self.working_label, 0, 0, Qt.AlignmentFlag.AlignHCenter)
-
-        self.status_label = QLabel('Starting...')
-        self.status_label.setFont(QFont('Helvetica', 16, QFont.Weight.Bold))
-        self.status_label.setStyleSheet('color: #d7d7f4; text-align: center;')
-        self.status_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        layout.addWidget(self.status_label, 1, 0, Qt.AlignmentFlag.AlignCenter)
-
-        self.progress_bar = QProgressBar()
-        self.progress_bar.setRange(1, progress_end)
-        self.progress_bar.setTextVisible(False)
-        self.progress_bar.setStyleSheet(
-            'QProgressBar {'
-                'border: 1px solid gray;'
-            '}'
-            'QProgressBar::chunk {'
-                'border: none;'
-                'background: #d7d7f4;'
-            '}'
-        )
-        layout.addWidget(self.progress_bar)
-
-        self.show()
-        QApplication.processEvents()
-
-    def change_text(self, text):
-        """
-        Method to change the text shown on the splash screen.
-
-        :param str text: The text to display.
-        """
-        self.status_label.setText(text)
-        self.progress_bar.setValue(self.progress_bar.value() + 1)
-        QApplication.processEvents()
-
-
-class CustomTextEdit(QTextEdit):
-    """
-    CustomTextEdit is an implementation of QTextEdit that adds spell-checking capabilities. Two different types of
-    spell checking are done depending on user input and current spell-check state: check previous word, or check
-    whole text. Also modifies the standard context menu to include spell check suggestions and an option to remove
-    extra white space from the text.
-
-    :param GUI gui: The GUI object
-    """
-    def __init__(self, gui):
-        super().__init__()
-        self.setObjectName('custom_text_edit')
-        self.gui = gui
-        self.full_spell_check_done = False
-        self.document().setDefaultStyleSheet(
-            'p {'
-                'font-family: ' + self.gui.main.user_settings['font_family'] + ';'
-                'font-size: ' + self.gui.main.user_settings['font_size'] + 'pt;'
-                'line-height: ' + self.gui.main.user_settings['line_spacing'] + ';'
-            '}'
-        )
-
-        self.textChanged.connect(self.text_changed)
-        self.word_spell_check = SpellCheck(self, 'previous', self.gui)
-        self.word_spell_check.setAutoDelete(False)
-
-    def keyReleaseEvent(self, evt):
-        cursor = self.textCursor()
-        if cursor.charFormat().foreground() == Qt.GlobalColor.red:
-            self.word_spell_check.type = 'current'
-            self.gui.main.spell_check_thread_pool.start(self.word_spell_check)
-        elif (evt.key() == Qt.Key.Key_Space
-                or evt.key() == Qt.Key.Key_Return
-                or evt.key() == Qt.Key.Key_Enter):
-            if not self.gui.main.user_settings['disable_spell_check']:
-                self.word_spell_check.type = 'previous'
-                self.gui.main.spell_check_thread_pool.start(self.word_spell_check)
-        else:
-            super().keyReleaseEvent(evt)
-
-    def keyPressEvent(self, evt):
-        if evt.key() == Qt.Key.Key_Enter or evt.key() == Qt.Key.Key_Return:
-            self.insertPlainText('\n')
-        else:
-            super().keyPressEvent(evt)
-
-    def contextMenuEvent(self, evt):
-        """
-        @override
-        Alters the standard context menu of this QTextEdit to include a list of spell-check words as well as an option
-        to remove extraneous whitespace from the text.
-        """
-        menu = self.createStandardContextMenu()
-
-        clean_whitespace_action = QAction("Remove extra whitespace")
-        clean_whitespace_action.triggered.connect(self.clean_whitespace)
-        menu.insertAction(menu.actions()[0], clean_whitespace_action)
-        menu.insertSeparator(menu.actions()[1])
-
-        if not self.gui.main.disable_spell_check:
-            cursor = self.cursorForPosition(evt.pos())
-            cursor.select(QTextCursor.SelectionType.WordUnderCursor)
-            word = cursor.selection().toPlainText()
-
-            if len(word) > 0:
-                upper = False
-                if word[0].isupper():
-                    upper = True
-
-                spell_check = SpellCheck(None, None, self.gui)
-                suggestions = spell_check.check_single_word(word)
-
-                next_menu_index = 0
-                spell_actions = {}
-
-                number_of_suggestions = len(suggestions)
-                if number_of_suggestions > 10: number_of_suggestions = 11
-
-                for i in range(number_of_suggestions):
-                    term = suggestions[i].term
-                    if upper:
-                        term = term[0].upper() + term[1:]
-                    spell_actions['action% s' % str(i)] = QAction(term)
-                    spell_actions['action% s' % str(i)].setData((cursor, term))
-                    spell_actions['action% s' % str(i)].triggered.connect(self.replace_word)
-                    menu.insertAction(menu.actions()[i], spell_actions['action% s' % str(i)])
-
-                    next_menu_index = i + 1
-
-                menu.insertSeparator(menu.actions()[next_menu_index])
-                action = QAction('Add to dictionary')
-                action.triggered.connect(lambda: self.gui.main.add_to_dictionary(self, spell_check.clean_word(word)))
-                menu.insertAction(menu.actions()[next_menu_index + 2], action)
-                menu.insertSeparator(menu.actions()[next_menu_index + 3])
-
-            menu.exec(evt.globalPos())
-            menu.close()
-
-    def toSimplifiedHtml(self):
-        """
-        Method to strip unneeded html tags and convert others to simpler tags
-        """
-        string = self.toHtml()
-        string = re.split('<body.*?>', string)[1]
-
-        # preserve any desired tags from getting removed during the wholesale <.*?> removal
-        string = re.sub('<p.*?>', '{p}', string)
-        string = re.sub('</p>', '{/p}', string)
-        string = re.sub('<ul.*?>', '{ul}', string)
-        string = re.sub('</ul>', '{/ul}', string)
-        string = re.sub('<li.*?>', '{li}', string)
-        string = re.sub('</li>', '{/li}', string)
-
-        bold_texts = re.findall('<span.*?font-weight.*?</span>', string)
-        for text in bold_texts:
-            new_text = re.sub('<.*?>', '', text)
-            # keep spaces on the outside of formatting marks
-            if new_text.startswith(' '):
-                new_text = ' {b}' + new_text[1:]
-            else:
-                new_text = '{b}' + new_text
-            if new_text.endswith(' '):
-                new_text = new_text[:-1] + '{/b} '
-            else:
-                new_text = new_text + '{/b}'
-            string = string.replace(text, new_text)
-
-        italic_texts = re.findall('<span.*?font-style.*?</span>', string)
-        for text in italic_texts:
-            new_text = re.sub('<.*?>', '', text)
-            # keep spaces on the outside of formatting marks
-            if new_text.startswith(' '):
-                new_text = ' {i}' + new_text[1:]
-            else:
-                new_text = '{i}' + new_text
-            if new_text.endswith(' '):
-                new_text = new_text[:-1] + '{/i} '
-            else:
-                new_text = new_text + '{/i}'
-            string = string.replace(text, new_text)
-
-        underline_texts = re.findall('<span.*?text-decoration.*?</span>', string)
-        for text in underline_texts:
-            new_text = re.sub('<.*?>', '', text)
-            # keep spaces on the outside of formatting marks
-            if new_text.startswith(' '):
-                new_text = ' {u}' + new_text[1:]
-            else:
-                new_text = '{u}' + new_text
-            if new_text.endswith(' '):
-                new_text = new_text[:-1] + '{/u} '
-            else:
-                new_text = new_text + '{/u}'
-            string = string.replace(text, new_text)
-
-        # convert preserved tags back to their original form
-        string = re.sub('<.*?>', '', string)
-        string = string.strip().replace('{p}', '<p>')
-        string = string.strip().replace('{/p}', '</p>\n')
-        string = string.replace('{ul}', '<ul>')
-        string = string.replace('{/ul}', '</ul>')
-        string = string.replace('{li}', '<li>')
-        string = string.replace('{/li}', '</li>\n')
-        string = string.replace('{b}', '<b>')
-        string = string.replace('{/b}', '</b>')
-        string = string.replace('{i}', '<i>')
-        string = string.replace('{/i}', '</i>')
-        string = string.replace('{u}', '<u>')
-        string = string.replace('{/u}', '</u>')
-
-        return string.strip()
-
-    def text_changed(self):
-        self.gui.changes = True
-
-    def replace_word(self):
-        """
-        Method to replace a misspelled word if the user chooses a replacement from the context menu.
-        """
-        sender = self.sender()
-        cursor = sender.data()[0]
-        term = sender.data()[1]
-
-        self.setTextCursor(cursor)
-        selected_word = self.textCursor().selectedText().strip()
-
-        # check if the original word includes punctuation so that it can be preserved
-        punctuation = ''
-        if not selected_word[len(selected_word) - 1].isalpha():
-            punctuation = selected_word[len(selected_word) - 1]
-
-        self.textCursor().removeSelectedText()
-        cursor = self.textCursor()
-        cursor.movePosition(QTextCursor.MoveOperation.NextCharacter, QTextCursor.MoveMode.KeepAnchor)
-        self.setTextCursor(cursor)
-
-        cursor = self.textCursor()
-        cursor.movePosition(QTextCursor.MoveOperation.PreviousCharacter, QTextCursor.MoveMode.MoveAnchor)
-        self.setTextCursor(cursor)
-
-        if len(punctuation) > 0:
-            self.textCursor().insertText(term + punctuation)
-        else:
-            self.textCursor().insertText(term + punctuation + ' ')
-
-        self.gui.changes = True
-
-    def clean_whitespace(self):
-        """
-        Method to remove duplicate spaces and tabs from the document.
-        """
-        component = self.gui.focusWidget()
-        if isinstance(component, QTextEdit):
-            string = component.toHtml()
-            string = re.sub(' +', ' ', string)
-            string = re.sub('\t+', '\t', string)
-
-            component.setHtml(string)
-        self.gui.changes = True
-
-    def paintEvent(self, evt):
-        if not self.full_spell_check_done:
-            spell_check = SpellCheck(self, 'whole', self.gui)
-            self.gui.main.spell_check_thread_pool.start(spell_check)
-            self.full_spell_check_done = True
-        super().paintEvent(evt)
-
-
-class ScriptureBox(QWidget):
-    """
-    Creates an independent QWidget that can be added or removed from layouts based on user's input.
-    """
-    def __init__(self):
-        super().__init__()
-        self.setObjectName('text_box')
-        self.setMaximumWidth(300)
-        text_layout = QVBoxLayout()
-        self.setLayout(text_layout)
-        text_title = QLabel()
-        text_title.setObjectName('text_title')
-        text_layout.addWidget(text_title)
-        self.text_edit = QTextEdit()
-        self.text_edit.setObjectName('text_box_text_edit')
-        self.text_edit.setReadOnly(True)
-        text_layout.addWidget(self.text_edit)
-        self.hide()
-
-    def clear(self):
-        self.text_edit.clear()
-
-
-class SearchBox(QWidget):
-    """
-    Creates an independent QWidget to be added to the main tabbed widget when the user performs a search.
-
-    :param GUI gui: The GUI object
-    """
-    def __init__(self, gui):
-        self.gui = gui
-        super().__init__()
-
-    def show_results(self, result_list):
-        """
-        Method to build the results widget.
-
-        :param list of str result_list: The list containing each list of results from the search.
-        """
-        results_widget_layout = QVBoxLayout()
-        self.setLayout(results_widget_layout)
-
-        results_header = QWidget()
-        header_layout = QHBoxLayout()
-        results_header.setLayout(header_layout)
-
-        results_label = QLabel()
-        header_layout.addWidget(results_label)
-
-        close_button = QPushButton()
-        close_button.setIcon(QIcon('resources/svg/spCloseIcon.svg'))
-        close_button.setToolTip('Close the search tab')
-        close_button.pressed.connect(self.remove_self)
-        header_layout.addStretch()
-        header_layout.addWidget(close_button)
-
-        results_widget_layout.addWidget(results_header)
-
-        # count the number of times the search term(s) was/were found in the record so that they can be sorted
-        filtered_results = []
-        for line in result_list:
-            counter = 0
-            for item in line[0]:
-                counter += 1
-
-            words_found = str(line[1])
-            words_found = words_found.replace('[', '')
-            words_found = words_found.replace(']', '')
-            words_found = words_found.replace("\'", '')
-            filtered_results.append((
-                str(line[0][0]),
-                str(line[2]),
-                words_found,
-                line[0][3],
-                line[0][16],
-                line[0][17],
-                line[0][21][0:100] + '...'))
-
-        model = QStandardItemModel(len(filtered_results), 5)
-        for i in range(len(filtered_results)):
-            for n in range(len(filtered_results[i])):
-                item = QStandardItem(filtered_results[i][n])
-                item.setEditable(False)
-                model.setItem(i, n, item)
-        model.setHeaderData(0, Qt.Orientation.Horizontal, 'ID')
-        model.setHeaderData(1, Qt.Orientation.Horizontal, '# of\r\nMatches')
-        model.setHeaderData(2, Qt.Orientation.Horizontal, 'Word(s) Found')
-        model.setHeaderData(3, Qt.Orientation.Horizontal, 'Sermon Text')
-        model.setHeaderData(4, Qt.Orientation.Horizontal, 'Sermon Title')
-        model.setHeaderData(5, Qt.Orientation.Horizontal, 'Sermon Date')
-        model.setHeaderData(6, Qt.Orientation.Horizontal, 'Sermon Snippet')
-
-        results_table_view = QTableView()
-        results_table_view.setModel(model)
-        results_table_view.setColumnWidth(0, 30)
-        results_table_view.setColumnWidth(1, 60)
-        results_table_view.setColumnWidth(2, 150)
-        results_table_view.setColumnWidth(3, 200)
-        results_table_view.setColumnWidth(4, 200)
-        results_table_view.setColumnWidth(5, 100)
-        results_table_view.setColumnWidth(6, 500)
-        results_table_view.setShowGrid(False)
-        results_table_view.setSelectionBehavior(QTableView.SelectionBehavior.SelectRows)
-        results_table_view.doubleClicked.connect(
-            lambda: self.retrieve_selection(model, results_table_view.selectionModel().currentIndex().row()))
-        results_widget_layout.addWidget(results_table_view)
-
-        if len(filtered_results) == 1:
-            results_label.setText(str(len(
-                filtered_results)) + ' result found.\nDouble-click a result below to open it.')
-        else:
-            results_label.setText(str(len(
-                filtered_results)) + ' results found.\nDouble-click a result below to open it.')
-
-    def retrieve_selection(self, model, selection):
-        """
-        Method to pull up whichever record the user selects
-
-        :param QStandardItemModel model: The model applied to the results_table_view
-        :param int selection: The row number of the results_table_view that was double-clicked
-        """
-        # be sure to check for changes before pulling up the new record
-        goon = True
-        if self.gui.changes:
-            goon = self.gui.main.ask_save()
-        if goon:
-            id_ = model.index(selection, 0).data()
-            index = 0
-            for item in self.gui.main.ids:
-                if int(item) == int(id_):
-                    break
-                index += 1
-            self.gui.main.get_by_index(index)
-            self.gui.tabbed_frame.setCurrentWidget(self.gui.tabbed_frame.widget(0))
-
-    def remove_self(self):
-        """
-        Method to remove this widget's tab from the GUI's tabbed widget.
-        """
-        self.gui.tabbed_frame.removeTab(5)
-        self.gui.tabbed_frame.setCurrentWidget(self.gui.tabbed_frame.widget(0))
-        self.destroy()
-
-
-class SermonView(QWidget):
-    def __init__(self, gui, text):
-        super().__init__()
-        self.gui = gui
-        self.text = text
-        self.button_bar = QWidget()
-        self.next_page_button = QPushButton('next')
-        self.previous_page_button = QPushButton('previous')
-        self.sermon_label = QLabel()
-        self.num_pages = 0
-        self.current_page = 0
-        self.current_font = QFont(
-            self.gui.main.user_settings['font_family'], int(self.gui.main.user_settings['font_size']))
-        self.page_label = None
-
-        self.init_components()
-        self.showMaximized()
-        self.gui.main.app.processEvents()
-
-        self.page_width = self.width() - 40
-        self.page_height = self.height() - self.button_bar.height() - 40
-
-        self.page_pixmaps = []
-        self.create_document_pages()
-        self.set_page()
-
-        self.previous_page_button.setEnabled(False)
-        if len(self.page_pixmaps) < 2:
-            self.next_page_button.setEnabled(False)
-        self.add_page_buttons()
-
-    def init_components(self):
-        self.setParent(self.gui)
-        self.setWindowFlag(Qt.WindowType.Window)
-        self.setWindowTitle('Sermon Viewer')
-
-        layout = QVBoxLayout(self)
-
-        button_layout = QHBoxLayout(self.button_bar)
-        layout.addWidget(self.button_bar)
-
-        zoom_out_button = QPushButton()
-        zoom_out_button.setObjectName('zoom_out')
-        zoom_out_button.setIcon(QIcon('resources/svg/spZoomOut.svg'))
-        zoom_out_button.setIconSize(QSize(36, 36))
-        zoom_out_button.pressed.connect(self.zoom)
-        button_layout.addWidget(zoom_out_button)
-        button_layout.addSpacing(20)
-
-        zoom_in_button = QPushButton()
-        zoom_in_button.setObjectName('zoom_in')
-        zoom_in_button.setIcon(QIcon('resources/svg/spZoomIn.svg'))
-        zoom_in_button.setIconSize(QSize(36, 36))
-        zoom_in_button.pressed.connect(self.zoom)
-        button_layout.addWidget(zoom_in_button)
-        button_layout.addStretch()
-
-        self.page_label = QLabel()
-        self.page_label.setFont(self.gui.bold_font)
-        button_layout.addWidget(self.page_label)
-        button_layout.addStretch()
-
-        layout.addWidget(self.sermon_label)
-        layout.addStretch()
-
-    def add_page_buttons(self):
-        page_button_next_stylesheet = (
-            'QPushButton {'
-            '   border: none;'
-            '   background: transparent;'
-            '   color: #00000000;'
-            '}'
-            'QPushButton:hover {'
-            '   background-color: rgba(255, 255, 255, 150);'
-            '   background-image: url("resources/svg/spNextPageIcon.svg");'
-            '   background-repeat: no-repeat;'
-            '   background-position: center;'
-            '}'
-            'QPushButton:pressed {'
-            '   background-color: #ffffff;'
-            '   color: #000000;'
-            '}'
-        )
-        page_button_previous_stylesheet = (
-            'QPushButton {'
-            '   border: none;'
-            '   background: transparent;'
-            '   color: #00000000;'
-            '}'
-            'QPushButton:hover {'
-            '   background-color: rgba(255, 255, 255, 150);'
-            '   background-image: url("resources/svg/spPrevPageIcon.svg");'
-            '   background-repeat: no-repeat;'
-            '   background-position: center;'
-            '}'
-            'QPushButton:pressed {'
-            '   background-color: #ffffff;'
-            '   color: #000000;'
-            '}'
-        )
-
-        self.previous_page_button.setParent(self)
-        self.previous_page_button.setObjectName('previous')
-        self.previous_page_button.setGeometry(
-            0, self.height() - self.page_height - 20, 200, self.page_height)
-        self.previous_page_button.setStyleSheet(page_button_previous_stylesheet)
-        self.previous_page_button.pressed.connect(self.set_page)
-        self.previous_page_button.show()
-
-        self.next_page_button.setParent(self)
-        self.next_page_button.setObjectName('next')
-        self.next_page_button.setGeometry(self.width() - 200, self.height() - self.page_height - 20, 200, self.page_height)
-        self.next_page_button.setStyleSheet(page_button_next_stylesheet)
-        self.next_page_button.pressed.connect(self.set_page)
-        self.next_page_button.show()
-
-    def create_document_pages(self):
-        document = QTextDocument()
-        document.setPageSize(QSizeF(self.page_width, self.page_height))
-        document.setDefaultFont(self.current_font)
-
-        text_option = QTextOption()
-        text_option.setWrapMode(QTextOption.WrapMode.WordWrap)
-        document.setDefaultTextOption(text_option)
-
-        document.setHtml(self.text)
-
-        self.page_pixmaps = []
-        current_y = 0
-        painter = QPainter()
-        self.num_pages = document.pageCount()
-        for i in range(document.pageCount()):
-            pixmap = QPixmap(self.page_width, self.page_height)
-            pixmap.fill(Qt.GlobalColor.white)
-
-            painter.begin(pixmap)
-            painter.translate(0, -current_y)
-            document.drawContents(painter)
-            painter.end()
-
-            self.page_pixmaps.append(pixmap)
-            current_y += self.page_height
-
-    def zoom(self):
-        if self.sender().objectName() == 'zoom_in':
-            self.current_font.setPointSize(self.current_font.pointSize() + 2)
-        else:
-            self.current_font.setPointSize(self.current_font.pointSize() - 2)
-        self.create_document_pages()
-        self.set_page()
-
-    def set_page(self):
-        if self.sender().objectName() == 'next':
-            self.current_page += 1
-        elif self.sender().objectName() == 'previous':
-            self.current_page -= 1
-
-        if self.current_page == 0:
-            self.previous_page_button.setEnabled(False)
-        else:
-            self.previous_page_button.setEnabled(True)
-
-        if self.current_page >= len(self.page_pixmaps) - 1:
-            self.current_page = len(self.page_pixmaps) - 1
-            self.next_page_button.setEnabled(False)
-        else:
-            self.next_page_button.setEnabled(True)
-
-        self.sermon_label.setPixmap(self.page_pixmaps[self.current_page])
-        self.page_label.setText(f'Page {self.current_page + 1} of {self.num_pages}')
