+++ widgets.py
@@ -6,22 +6,421 @@ import sys
 from os.path import exists
 
 import wmi
-from PyQt6.QtCore import Qt, QSizeF, QRectF
-from PyQt6.QtGui import QStandardItemModel, QFontDatabase, QStandardItem, QPixmap, QTextCursor, QFont, QIcon, \
-    QTextDocument, QTextOption, QPainter
+from PyQt6.QtCore import pyqtSignal, Qt, QSize, QSizeF, QRectF
+from PyQt6.QtGui import QPixmap, QFont, QAction, QTextCursor, QIcon, QStandardItemModel, QStandardItem, QTextDocument, \
+    QTextOption, QPainter, QTextListFormat, QTextCharFormat, QFontDatabase
 from PyQt6.QtPrintSupport import QPrinter
-from PyQt6.QtWidgets import QFileDialog, QWidget, QVBoxLayout, QLabel, QTableView, QPushButton, \
-    QTabWidget, QHBoxLayout, QComboBox, QTextBrowser, QLineEdit, QTextEdit, QDateEdit, QMessageBox, QSpinBox
+from PyQt6.QtWidgets import QDialog, QTextEdit, QWidget, QGridLayout, QLabel, QProgressBar, QApplication, QVBoxLayout, \
+    QHBoxLayout, QPushButton, QTableView, QMessageBox, QLineEdit, QComboBox, QFileDialog, QTabWidget, QTextBrowser, \
+    QSpinBox, QDateEdit
 from pynput.keyboard import Key, Controller
 
-from runnables import LoadDictionary, SpellCheck
-from toolbar import Toolbar
+
+class StartupSplash(QDialog):
+    update_text = pyqtSignal(str)
+    end = pyqtSignal()
+
+    def __init__(self, gui, progress_end):
+        super().__init__()
+        self.gui = gui
+        self.update_text.connect(self.change_text)
+        self.end.connect(lambda: self.done(0))
+
+        self.setWindowFlag(Qt.WindowType.FramelessWindowHint)
+        self.setModal(True)
+        self.setAttribute(Qt.WidgetAttribute.WA_TranslucentBackground)
+        self.setMinimumWidth(300)
+
+        layout = QGridLayout()
+        self.setLayout(layout)
+
+        self.working_label = QLabel()
+        self.working_label.setAutoFillBackground(False)
+        self.working_label.setPixmap(QPixmap('resources/icon.png'))
+        layout.addWidget(self.working_label, 0, 0, Qt.AlignmentFlag.AlignHCenter)
+
+        self.status_label = QLabel('Starting...')
+        self.status_label.setFont(QFont('Helvetica', 16, QFont.Weight.Bold))
+        self.status_label.setStyleSheet('color: #d7d7f4; text-align: center;')
+        self.status_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        layout.addWidget(self.status_label, 1, 0, Qt.AlignmentFlag.AlignCenter)
+
+        self.progress_bar = QProgressBar()
+        self.progress_bar.setRange(1, progress_end)
+        self.progress_bar.setTextVisible(False)
+        self.progress_bar.setStyleSheet(
+            'QProgressBar {'
+                'border: 1px solid gray;'
+            '}'
+            'QProgressBar::chunk {'
+                'border: none;'
+                'background: #d7d7f4;'
+            '}'
+        )
+        layout.addWidget(self.progress_bar)
+
+        self.show()
+        QApplication.processEvents()
+
+    def change_text(self, text):
+        """
+        Method to change the text shown on the splash screen.
+
+        :param str text: The text to display.
+        """
+        self.status_label.setText(text)
+        self.progress_bar.setValue(self.progress_bar.value() + 1)
+        QApplication.processEvents()
+
+
+class Toolbar(QWidget):
+    """
+    Toolbar creates the uppermost QWidget of the GUI that holds formatting, search, and navigation elements.
+    """
+    def __init__(self, gui, main):
+        super().__init__()
+        self.gui = gui
+        self.main = main
+        self.setObjectName('toolbar')
+
+        icon_size = QSize(16, 16)
+
+        toolbar_layout = QHBoxLayout(self)
+        toolbar_layout.setContentsMargins(5, 5, 5, 5)
+
+        self.undo_button = QPushButton()
+        self.undo_button.setIcon(QIcon('resources/svg/spUndoIcon.svg'))
+        self.undo_button.setIconSize(icon_size)
+        self.undo_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.undo_button.clicked.connect(self.gui.menu_bar.press_ctrl_z)
+        self.undo_button.setToolTip('Undo')
+        toolbar_layout.addWidget(self.undo_button)
+
+        self.redo_button = QPushButton()
+        self.redo_button.setIcon(QIcon('resources/svg/spRedoIcon.svg'))
+        self.redo_button.setIconSize(icon_size)
+        self.redo_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.redo_button.clicked.connect(self.gui.menu_bar.press_ctrl_y)
+        self.redo_button.setToolTip('Redo')
+        toolbar_layout.addWidget(self.redo_button)
+        toolbar_layout.addSpacing(20)
+
+        self.bold_button = QPushButton()
+        self.bold_button.setCheckable(True)
+        self.bold_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.bold_button.clicked.connect(self.set_bold)
+        self.bold_button.setIcon(QIcon('resources/svg/spBoldIcon.svg'))
+        self.bold_button.setIconSize(icon_size)
+        self.bold_button.setToolTip('Bold\n(Ctrl+B)')
+        toolbar_layout.addWidget(self.bold_button)
+
+        self.italic_button = QPushButton()
+        self.italic_button.setCheckable(True)
+        self.italic_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.italic_button.clicked.connect(self.set_italic)
+        self.italic_button.setIcon(QIcon('resources/svg/spItalicIcon.svg'))
+        self.italic_button.setIconSize(icon_size)
+        self.italic_button.setToolTip('Italic\n(Ctrl+I)')
+        toolbar_layout.addWidget(self.italic_button)
+
+        self.underline_button = QPushButton()
+        self.underline_button.setCheckable(True)
+        self.underline_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.underline_button.clicked.connect(self.set_underline)
+        self.underline_button.setIcon(QIcon('resources/svg/spUnderlineIcon.svg'))
+        self.underline_button.setIconSize(icon_size)
+        self.underline_button.setToolTip('Underline\n(Ctrl+U)')
+        toolbar_layout.addWidget(self.underline_button)
+
+        self.bullet_button = QPushButton()
+        self.bullet_button.setCheckable(True)
+        self.bullet_button.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.bullet_button.clicked.connect(self.set_bullet)
+        self.bullet_button.setIcon(QIcon('resources/svg/spBulletIcon.svg'))
+        self.bullet_button.setIconSize(icon_size)
+        self.bullet_button.setToolTip('Add Bullets\n(Ctrl+Shift+B)')
+        toolbar_layout.addWidget(self.bullet_button)
+        toolbar_layout.addSpacing(20)
+
+        self.text_visible = QPushButton()
+        self.text_visible.setObjectName('text_visible')
+        self.text_visible.setCheckable(True)
+        self.text_visible.setFocusPolicy(Qt.FocusPolicy.NoFocus)
+        self.text_visible.setToolTip('Show Sermon Text on All Tabs')
+        self.text_visible.setIcon(QIcon('resources/svg/spShowText.svg'))
+        self.text_visible.setIconSize(QSize(round(icon_size.width() * 2.5), icon_size.height()))
+        self.text_visible.clicked.connect(self.keep_text_visible)
+        toolbar_layout.addWidget(self.text_visible)
+
+        toolbar_layout.addStretch(1)
+
+        dates_label = QLabel('Choose Record by Date:')
+        dates_label.setAutoFillBackground(False)
+        toolbar_layout.addWidget(dates_label)
+
+        self.dates_cb = QComboBox()
+        self.dates_cb.addItems(self.main.dates)
+        self.dates_cb.currentIndexChanged.connect(lambda: self.main.get_by_index(self.dates_cb.currentIndex()))
+        self.dates_cb.setMinimumWidth(100)
+        toolbar_layout.addWidget(self.dates_cb)
+
+        references_label = QLabel('Choose Record by Sermon Reference:')
+        references_label.setAutoFillBackground(False)
+        toolbar_layout.addWidget(references_label)
+
+        self.references_cb = QComboBox()
+        for item in self.main.references:
+            self.references_cb.addItem(item[0])
+        self.references_cb.currentIndexChanged.connect(
+            lambda: self.get_index_of_reference(self.references_cb.currentIndex()))
+        toolbar_layout.addWidget(self.references_cb)
+
+        search_label = QLabel('Keyword Search:')
+        search_label.setAutoFillBackground(False)
+        toolbar_layout.addWidget(search_label)
+
+        search_field = QLineEdit()
+        search_field.setMinimumWidth(200)
+        search_field.returnPressed.connect(lambda: self.do_search(search_field.text()))
+        toolbar_layout.addWidget(search_field)
+
+        toolbar_layout.addStretch(1)
+
+        self.first_rec_button = QPushButton()
+        self.first_rec_button.setIcon(QIcon('resources/svg/spFirstRecIcon.svg'))
+        self.first_rec_button.setIconSize(icon_size)
+        self.first_rec_button.clicked.connect(self.main.first_rec)
+        self.first_rec_button.setToolTip('Jump to First Record')
+        toolbar_layout.addWidget(self.first_rec_button)
+
+        self.prev_rec_button = QPushButton()
+        self.prev_rec_button.setIcon(QIcon('resources/svg/spPrevRecIcon.svg'))
+        self.prev_rec_button.setIconSize(icon_size)
+        self.prev_rec_button.clicked.connect(self.main.prev_rec)
+        self.prev_rec_button.setToolTip('Go to Previous Record')
+        toolbar_layout.addWidget(self.prev_rec_button)
+
+        self.next_rec_button = QPushButton()
+        self.next_rec_button.setIcon(QIcon('resources/svg/spNextRecIcon.svg'))
+        self.next_rec_button.clicked.connect(self.main.next_rec)
+        self.next_rec_button.setToolTip('Go to Next Record')
+        toolbar_layout.addWidget(self.next_rec_button)
+
+        self.last_rec_button = QPushButton()
+        self.last_rec_button.setIcon(QIcon('resources/svg/spLastRecIcon.svg'))
+        self.last_rec_button.clicked.connect(self.main.last_rec)
+        self.last_rec_button.setToolTip('Jump to Last Record')
+        toolbar_layout.addWidget(self.last_rec_button)
+
+        self.new_rec_button = QPushButton()
+        self.new_rec_button.setIcon(QIcon('resources/svg/spNewIcon.svg'))
+        self.new_rec_button.clicked.connect(self.main.new_rec)
+        self.new_rec_button.setToolTip('Create a New Record')
+        toolbar_layout.addWidget(self.new_rec_button)
+        toolbar_layout.addSpacing(20)
+
+        self.save_button = QPushButton()
+        self.save_button.setIcon(QIcon('resources/svg/spSaveIcon.svg'))
+        self.save_button.setIconSize(icon_size)
+        self.save_button.clicked.connect(self.main.save_rec)
+        self.save_button.setToolTip('Save this Record')
+        toolbar_layout.addWidget(self.save_button)
+
+        self.print_button = QPushButton()
+        self.print_button.setIcon(QIcon('resources/svg/spPrintIcon.svg'))
+        self.print_button.setIconSize(icon_size)
+        self.print_button.clicked.connect(self.gui.menu_bar.print_rec)
+        self.print_button.setToolTip('Print this Record')
+        toolbar_layout.addWidget(self.print_button)
+
+        self.id_label = QLabel()
+        toolbar_layout.addWidget(self.id_label)
+
+    def keep_text_visible(self):
+        """
+        Handle the user's toggling of the text_visible button.
+        """
+        check_state = self.text_visible.isChecked()
+
+        if check_state:
+            # add the reference and passage text to each tab's text_box then make it show
+            num_tabs = self.gui.tabbed_frame.count()
+            for i in range(num_tabs):
+                if i > 0:
+                    frame = self.gui.tabbed_frame.widget(i)
+                    widget = frame.findChild(QWidget, 'text_box')
+                    text_title = widget.findChild(QLabel, 'text_title')
+                    text_edit = widget.findChild(QTextEdit, 'text_box_text_edit')
+                    text_edit.setText(self.gui.sermon_text_edit.toPlainText())
+                    text_title.setText(self.gui.sermon_reference_field.text())
+                    if widget:
+                        widget.show()
+        else:
+            # hide the text_box on each tab
+            num_tabs = self.gui.tabbed_frame.count()
+            for i in range(num_tabs):
+                if i > 0:
+                    frame = self.gui.tabbed_frame.widget(i)
+                    widget = frame.findChild(QWidget, 'text_box')
+                    if widget:
+                        widget.hide()
+
+    def get_index_of_reference(self, index):
+        """
+        Method to find the index number of the user's chosen reference.
+
+        :param int index: The index of the reference combo box's chosen reference
+        """
+        id_to_find = self.main.references[index][1]
+        counter = 0
+        for item in self.main.ids:
+            if item == id_to_find:
+                break
+            counter += 1
+        self.main.get_by_index(counter)
+
+    def do_search(self, text):
+        """
+        Method to call get_search_results from SermonPrepDatabase and display the results
+
+        :param str text: The user's search term(s)
+        """
+        result_list = self.main.get_search_results(text)
+        if len(result_list) == 0:
+            QMessageBox.information(
+                None,
+                'No Results',
+                'No results were found. Please try your search again.',
+                QMessageBox.StandardButton.Ok
+            )
+        else:
+            from gui import SearchBox
+            search_box = SearchBox(self.gui)
+            self.gui.tabbed_frame.addTab(search_box, QIcon('resources/svg/spSearchIcon.svg'), 'Search')
+            search_box.show_results(result_list)
+            self.gui.tabbed_frame.setCurrentWidget(search_box)
+
+    def set_bold(self):
+        """
+        Method to toggle the bold state of the text at the user's cursor or selection.
+        """
+        component = self.gui.focusWidget()
+        if isinstance(component, QTextEdit):
+            cursor = component.textCursor()
+            # handle this differently if the user has a section of text selected
+            if cursor.hasSelection():
+                selection_start = cursor.selectionStart()
+                selection_end = cursor.selectionEnd()
+                cursor.setPosition(selection_start, QTextCursor.MoveMode.MoveAnchor)
+                cursor.setPosition(selection_end, QTextCursor.MoveMode.KeepAnchor)
+                char_format = cursor.charFormat()
+                if char_format.font().bold():
+                    char_format.setFontWeight(QFont.Weight.Normal)
+                    cursor.mergeCharFormat(char_format)
+                else:
+                    char_format.setFontWeight(QFont.Weight.Bold)
+                    cursor.mergeCharFormat(char_format)
+            else:
+                font = cursor.charFormat().font()
+                if font.weight() == QFont.Weight.Normal:
+                    font.setWeight(QFont.Weight.Bold)
+                    component.setCurrentFont(font)
+                else:
+                    font.setWeight(QFont.Weight.Normal)
+                    component.setCurrentFont(font)
+        else:
+            self.bold_button.setChecked(False)
+
+    def set_italic(self):
+        """
+        Method to toggle the italic state of the text at the user's cursor or selection.
+        """
+        component = self.gui.focusWidget()
+        if isinstance(component, QTextEdit):
+            cursor = component.textCursor()
+            if cursor.hasSelection():
+                font = QTextCharFormat(cursor.charFormat())
+                if not font.fontItalic():
+                    font.setFontItalic(True)
+                    cursor.setCharFormat(QTextCharFormat(font))
+                else:
+                    font.setFontItalic(False)
+                    cursor.mergeCharFormat(QTextCharFormat(font))
+            else:
+                font = cursor.charFormat().font()
+                if not font.italic():
+                    font.setItalic(True)
+                    component.setCurrentFont(font)
+                else:
+                    font.setItalic(False)
+                    component.setCurrentFont(font)
+        else:
+            self.italic_button.setChecked(False)
+
+    def set_underline(self):
+        """
+        Method to toggle the underline state of the text at the user's cursor or selection.
+        """
+        component = self.gui.focusWidget()
+        if isinstance(component, QTextEdit):
+            cursor = component.textCursor()
+            if cursor.hasSelection():
+                font = QTextCharFormat(cursor.charFormat())
+                if not font.fontUnderline():
+                    font.setFontUnderline(True)
+                    cursor.setCharFormat(QTextCharFormat(font))
+                else:
+                    font.setFontUnderline(False)
+                    cursor.mergeCharFormat(QTextCharFormat(font))
+            else:
+                font = cursor.charFormat().font()
+                if not font.underline():
+                    font.setUnderline(True)
+                    component.setCurrentFont(font)
+                else:
+                    font.setUnderline(False)
+                    component.setCurrentFont(font)
+        else:
+            self.underline_button.setChecked(False)
+
+    def set_bullet(self):
+        """
+        Method to toggle the bulleted list state of the text at the user's cursor or selection.
+        """
+        component = self.gui.focusWidget()
+        if isinstance(component, QTextEdit):
+            cursor = component.textCursor()
+            text_list = cursor.currentList()
+            if text_list:
+                start = cursor.selectionStart()
+                end = cursor.selectionEnd()
+                removed = 0
+                for i in range(text_list.count()):
+                    item = text_list.item(i - removed)
+                    if (item.position() <= end and
+                            item.position() + item.length() > start):
+                        text_list.remove(item)
+                        block_cursor = QTextCursor(item)
+                        block_format = block_cursor.blockFormat()
+                        block_format.setIndent(0)
+                        block_cursor.mergeBlockFormat(block_format)
+                        removed += 1
+                component.setTextCursor(cursor)
+            else:
+                list_format = QTextListFormat()
+                style = QTextListFormat.Style.ListDisc
+                list_format.setStyle(style)
+                cursor.createList(list_format)
+                component.setTextCursor(cursor)
+        else:
+            self.bullet_button.setChecked(False)
 
 
 class MenuBar:
     """
     Builds the QMainWindow's menuBar and handles the actions performed by the menu.
     """
+
     def __init__(self, gui, main):
         """
         :param GUI gui: The program's GUI Object
@@ -333,8 +732,10 @@ class MenuBar:
                     QMessageBox.warning(
                         self.gui,
                         'Bad Format',
-                        'There is a problem with your XML bible: ' + file[0] + '. Try downloading it again or ensuring '
-                        'that it is formatted according to Zefania standards.',
+                        'There is a problem with your XML bible: '
+                            + file[0]
+                            + '. Try downloading it again or ensuring '
+                            'that it is formatted according to Zefania standards.',
                         QMessageBox.StandardButton.Ok
                     )
 
@@ -382,7 +783,7 @@ class MenuBar:
         self.rename_widget.setLayout(rename_widget_layout)
 
         rename_label = QLabel('Use this table to set new names for any of the labels in this program.\n'
-                             'Double-click a label under "New Label" to rename it')
+                              'Double-click a label under "New Label" to rename it')
         rename_widget_layout.addWidget(rename_label)
 
         # provide two columns in the QTableView; one that will contain the labels as they are, the other to provide
@@ -539,6 +940,7 @@ class MenuBar:
         else:
             # if spell check is enabled after having been disabled at startup, the dictionary will need to be loaded
             if not self.main.sym_spell:
+                from runnables import LoadDictionary
                 ld = LoadDictionary(self.main)
                 self.main.load_dictionary_thread_pool.start(ld)
             self.main.user_settings['disable_spell_check'] = False
@@ -546,6 +948,7 @@ class MenuBar:
 
             # run spell check on all CustomTextEdits
             for widget in self.gui.tabbed_frame.findChildren(QTextEdit, 'custom_text_edit'):
+                from runnables import SpellCheck
                 spell_check = SpellCheck(widget, 'whole', self.gui)
                 self.gui.main.spell_check_thread_pool.start(spell_check)
 
@@ -577,7 +980,7 @@ class MenuBar:
         about_layout = QVBoxLayout()
         self.about_win.setLayout(about_layout)
 
-        about_label = QLabel('Sermon Prep Database v.5.0.3.001')
+        about_label = QLabel('Sermon Prep Database v.5.0.3.002')
         about_layout.addWidget(about_label)
 
         about_text = QTextBrowser()
@@ -586,19 +989,19 @@ class MenuBar:
             modify it under the terms of the GNU General Public License (GNU GPL)
             published by the Free Software Foundation, either version 3 of the
             License, or (at your option) any later version.<br><br>
-    
+
             This program is distributed in the hope that it will be useful,
             but WITHOUT ANY WARRANTY; without even the implied warranty of
             MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
             GNU General Public License for more details.<br><br>
-    
+
             You should have received a copy of the GNU General Public License
             along with this program.  If not, see <a href="http://www.gnu.org/licenses/">http://www.gnu.org/licenses/</a>.<br><br>
-    
+
             The Sermon Prep Database program includes Artifex Software's GhostScript,
             licensed under the GNU Affero General Public License (GNU AGPL). See
             <a href="https://www.ghostscript.com/licensing/index.html">https://www.ghostscript.com/licensing/index.html</a> for more information.<br><br>
-    
+
             This program is a work-in-progress by a guy who is not, in no way, a
             professional programmer. If you run into any problems, unexpected behavior,
             missing features, or attempts to assimilate your unique biological and
@@ -656,7 +1059,8 @@ class MenuBar:
 
         ok_button = QPushButton('OK')
         ok_button.clicked.connect(
-            lambda: self.gui.apply_font(family_combo_box.currentText(), size_combo_box.currentText(), font_chooser, True))
+            lambda: self.gui.apply_font(family_combo_box.currentText(), size_combo_box.currentText(), font_chooser,
+                                        True))
         bottom_layout.addWidget(ok_button)
 
         cancel_button = QPushButton('Cancel')
@@ -740,6 +1144,7 @@ class ShowHelp(QTabWidget):
     """
     Method to create a QTabbedWidget to hold the different help topics.
     """
+
     def __init__(self, gui, spd):
         super().__init__()
         self.gui = gui
@@ -1024,8 +1429,8 @@ class ShowHelp(QTabWidget):
 
         research_text = QTextEdit()
         research_text.setHtml('In the research tab, you can jot down notes as you do research on the text for your '
-                               'sermon.<br><br>You are able to use basic formatting as you record your notes: bold, '
-                               'underline, italic, and bullet points.')
+                              'sermon.<br><br>You are able to use basic formatting as you record your notes: bold, '
+                              'underline, italic, and bullet points.')
         research_text.setFont(self.gui.standard_font)
         research_text.setReadOnly(True)
         research_layout.addWidget(research_text)
@@ -1057,13 +1462,14 @@ class ShowHelp(QTabWidget):
         sermon_layout.addWidget(sermon_text)
 
         self.addTab(sermon_widget, 'Sermon Tab')
-        
-        
+
+
 class FontFaceComboBox(QComboBox):
     """
     Creates a custom QComboBox that displays all fonts on the system in their own style.
     :param gui.GUI gui: The current instance of GUI
     """
+
     def __init__(self, gui):
         """
         :param gui.GUI gui: The current instance of GUI
@@ -1094,6 +1500,7 @@ class PrintHandler(QWidget):
 
     :param GUI gui: the current instance of GUI
     """
+
     def __init__(self, gui):
         super().__init__()
         self.gui = gui
@@ -1193,9 +1600,10 @@ class PrintHandler(QWidget):
         document.setDefaultStyleSheet(
             'p { '
             'font-family: "' + self.print_font.family() + '";'
-            'font-size: ' + str(self.print_font.pointSize()) + 'pt;'
-            'line-height: ' + str(self.line_height) + ';'
-            '}'
+                                                          'font-size: ' + str(self.print_font.pointSize()) + 'pt;'
+                                                                                                             'line-height: ' + str(
+                self.line_height) + ';'
+                                    '}'
         )
         document.setHtml(self.html)
 
@@ -1210,7 +1618,7 @@ class PrintHandler(QWidget):
             pixmap.fill(Qt.GlobalColor.white)
 
             painter.begin(pixmap)
-            painter.translate(0, -current_y) # translates the pixmap to the y position of the current page
+            painter.translate(0, -current_y)  # translates the pixmap to the y position of the current page
             document.drawContents(painter)
             painter.end()
 
@@ -1398,3 +1806,550 @@ class PrintHandler(QWidget):
         """
         self.document.print(self.printer)
         self.deleteLater()
+
+
+class CustomTextEdit(QTextEdit):
+    """
+    CustomTextEdit is an implementation of QTextEdit that adds spell-checking capabilities. Two different types of
+    spell checking are done depending on user input and current spell-check state: check previous word, or check
+    whole text. Also modifies the standard context menu to include spell check suggestions and an option to remove
+    extra white space from the text.
+
+    :param GUI gui: The GUI object
+    """
+    def __init__(self, gui):
+        super().__init__()
+        self.setObjectName('custom_text_edit')
+        self.gui = gui
+        self.full_spell_check_done = False
+        self.document().setDefaultStyleSheet(
+            'p {'
+                'font-family: ' + self.gui.main.user_settings['font_family'] + ';'
+                'font-size: ' + self.gui.main.user_settings['font_size'] + 'pt;'
+                'line-height: ' + self.gui.main.user_settings['line_spacing'] + ';'
+            '}'
+        )
+
+        self.textChanged.connect(self.text_changed)
+        from runnables import SpellCheck
+        self.word_spell_check = SpellCheck(self, 'previous', self.gui)
+        self.word_spell_check.setAutoDelete(False)
+
+    def keyReleaseEvent(self, evt):
+        cursor = self.textCursor()
+        if cursor.charFormat().foreground() == Qt.GlobalColor.red:
+            self.word_spell_check.type = 'current'
+            self.gui.main.spell_check_thread_pool.start(self.word_spell_check)
+        elif (evt.key() == Qt.Key.Key_Space
+                or evt.key() == Qt.Key.Key_Return
+                or evt.key() == Qt.Key.Key_Enter):
+            if not self.gui.main.user_settings['disable_spell_check']:
+                self.word_spell_check.type = 'previous'
+                self.gui.main.spell_check_thread_pool.start(self.word_spell_check)
+        else:
+            super().keyReleaseEvent(evt)
+
+    def keyPressEvent(self, evt):
+        if evt.key() == Qt.Key.Key_Enter or evt.key() == Qt.Key.Key_Return:
+            self.insertPlainText('\n')
+        else:
+            super().keyPressEvent(evt)
+
+    def contextMenuEvent(self, evt):
+        """
+        @override
+        Alters the standard context menu of this QTextEdit to include a list of spell-check words as well as an option
+        to remove extraneous whitespace from the text.
+        """
+        menu = self.createStandardContextMenu()
+
+        clean_whitespace_action = QAction("Remove extra whitespace")
+        clean_whitespace_action.triggered.connect(self.clean_whitespace)
+        menu.insertAction(menu.actions()[0], clean_whitespace_action)
+        menu.insertSeparator(menu.actions()[1])
+
+        if not self.gui.main.disable_spell_check:
+            cursor = self.cursorForPosition(evt.pos())
+            cursor.select(QTextCursor.SelectionType.WordUnderCursor)
+            word = cursor.selection().toPlainText()
+
+            if len(word) > 0:
+                upper = False
+                if word[0].isupper():
+                    upper = True
+
+                from runnables import SpellCheck
+                spell_check = SpellCheck(None, None, self.gui)
+                suggestions = spell_check.check_single_word(word)
+
+                next_menu_index = 0
+                spell_actions = {}
+
+                number_of_suggestions = len(suggestions)
+                if number_of_suggestions > 10: number_of_suggestions = 11
+
+                for i in range(number_of_suggestions):
+                    term = suggestions[i].term
+                    if upper:
+                        term = term[0].upper() + term[1:]
+                    spell_actions['action% s' % str(i)] = QAction(term)
+                    spell_actions['action% s' % str(i)].setData((cursor, term))
+                    spell_actions['action% s' % str(i)].triggered.connect(self.replace_word)
+                    menu.insertAction(menu.actions()[i], spell_actions['action% s' % str(i)])
+
+                    next_menu_index = i + 1
+
+                menu.insertSeparator(menu.actions()[next_menu_index])
+                action = QAction('Add to dictionary')
+                action.triggered.connect(lambda: self.gui.main.add_to_dictionary(self, spell_check.clean_word(word)))
+                menu.insertAction(menu.actions()[next_menu_index + 2], action)
+                menu.insertSeparator(menu.actions()[next_menu_index + 3])
+
+            menu.exec(evt.globalPos())
+            menu.close()
+
+    def toSimplifiedHtml(self):
+        """
+        Method to strip unneeded html tags and convert others to simpler tags
+        """
+        string = self.toHtml()
+        string = re.split('<body.*?>', string)[1]
+
+        # preserve any desired tags from getting removed during the wholesale <.*?> removal
+        string = re.sub('<p.*?>', '{p}', string)
+        string = re.sub('</p>', '{/p}', string)
+        string = re.sub('<ul.*?>', '{ul}', string)
+        string = re.sub('</ul>', '{/ul}', string)
+        string = re.sub('<li.*?>', '{li}', string)
+        string = re.sub('</li>', '{/li}', string)
+
+        bold_texts = re.findall('<span.*?font-weight.*?</span>', string)
+        for text in bold_texts:
+            new_text = re.sub('<.*?>', '', text)
+            # keep spaces on the outside of formatting marks
+            if new_text.startswith(' '):
+                new_text = ' {b}' + new_text[1:]
+            else:
+                new_text = '{b}' + new_text
+            if new_text.endswith(' '):
+                new_text = new_text[:-1] + '{/b} '
+            else:
+                new_text = new_text + '{/b}'
+            string = string.replace(text, new_text)
+
+        italic_texts = re.findall('<span.*?font-style.*?</span>', string)
+        for text in italic_texts:
+            new_text = re.sub('<.*?>', '', text)
+            # keep spaces on the outside of formatting marks
+            if new_text.startswith(' '):
+                new_text = ' {i}' + new_text[1:]
+            else:
+                new_text = '{i}' + new_text
+            if new_text.endswith(' '):
+                new_text = new_text[:-1] + '{/i} '
+            else:
+                new_text = new_text + '{/i}'
+            string = string.replace(text, new_text)
+
+        underline_texts = re.findall('<span.*?text-decoration.*?</span>', string)
+        for text in underline_texts:
+            new_text = re.sub('<.*?>', '', text)
+            # keep spaces on the outside of formatting marks
+            if new_text.startswith(' '):
+                new_text = ' {u}' + new_text[1:]
+            else:
+                new_text = '{u}' + new_text
+            if new_text.endswith(' '):
+                new_text = new_text[:-1] + '{/u} '
+            else:
+                new_text = new_text + '{/u}'
+            string = string.replace(text, new_text)
+
+        # convert preserved tags back to their original form
+        string = re.sub('<.*?>', '', string)
+        string = string.strip().replace('{p}', '<p>')
+        string = string.strip().replace('{/p}', '</p>\n')
+        string = string.replace('{ul}', '<ul>')
+        string = string.replace('{/ul}', '</ul>')
+        string = string.replace('{li}', '<li>')
+        string = string.replace('{/li}', '</li>\n')
+        string = string.replace('{b}', '<b>')
+        string = string.replace('{/b}', '</b>')
+        string = string.replace('{i}', '<i>')
+        string = string.replace('{/i}', '</i>')
+        string = string.replace('{u}', '<u>')
+        string = string.replace('{/u}', '</u>')
+
+        return string.strip()
+
+    def text_changed(self):
+        self.gui.changes = True
+
+    def replace_word(self):
+        """
+        Method to replace a misspelled word if the user chooses a replacement from the context menu.
+        """
+        sender = self.sender()
+        cursor = sender.data()[0]
+        term = sender.data()[1]
+
+        self.setTextCursor(cursor)
+        selected_word = self.textCursor().selectedText().strip()
+
+        # check if the original word includes punctuation so that it can be preserved
+        punctuation = ''
+        if not selected_word[len(selected_word) - 1].isalpha():
+            punctuation = selected_word[len(selected_word) - 1]
+
+        self.textCursor().removeSelectedText()
+        cursor = self.textCursor()
+        cursor.movePosition(QTextCursor.MoveOperation.NextCharacter, QTextCursor.MoveMode.KeepAnchor)
+        self.setTextCursor(cursor)
+
+        cursor = self.textCursor()
+        cursor.movePosition(QTextCursor.MoveOperation.PreviousCharacter, QTextCursor.MoveMode.MoveAnchor)
+        self.setTextCursor(cursor)
+
+        if len(punctuation) > 0:
+            self.textCursor().insertText(term + punctuation)
+        else:
+            self.textCursor().insertText(term + punctuation + ' ')
+
+        self.gui.changes = True
+
+    def clean_whitespace(self):
+        """
+        Method to remove duplicate spaces and tabs from the document.
+        """
+        component = self.gui.focusWidget()
+        if isinstance(component, QTextEdit):
+            string = component.toHtml()
+            string = re.sub(' +', ' ', string)
+            string = re.sub('\t+', '\t', string)
+
+            component.setHtml(string)
+        self.gui.changes = True
+
+    def paintEvent(self, evt):
+        if not self.full_spell_check_done:
+            from runnables import SpellCheck
+            spell_check = SpellCheck(self, 'whole', self.gui)
+            self.gui.main.spell_check_thread_pool.start(spell_check)
+            self.full_spell_check_done = True
+        super().paintEvent(evt)
+
+
+class ScriptureBox(QWidget):
+    """
+    Creates an independent QWidget that can be added or removed from layouts based on user's input.
+    """
+    def __init__(self):
+        super().__init__()
+        self.setObjectName('text_box')
+        self.setMaximumWidth(300)
+        text_layout = QVBoxLayout()
+        self.setLayout(text_layout)
+        text_title = QLabel()
+        text_title.setObjectName('text_title')
+        text_layout.addWidget(text_title)
+        self.text_edit = QTextEdit()
+        self.text_edit.setObjectName('text_box_text_edit')
+        self.text_edit.setReadOnly(True)
+        text_layout.addWidget(self.text_edit)
+        self.hide()
+
+    def clear(self):
+        self.text_edit.clear()
+
+
+class SearchBox(QWidget):
+    """
+    Creates an independent QWidget to be added to the main tabbed widget when the user performs a search.
+
+    :param GUI gui: The GUI object
+    """
+    def __init__(self, gui):
+        self.gui = gui
+        super().__init__()
+
+    def show_results(self, result_list):
+        """
+        Method to build the results widget.
+
+        :param list of str result_list: The list containing each list of results from the search.
+        """
+        results_widget_layout = QVBoxLayout()
+        self.setLayout(results_widget_layout)
+
+        results_header = QWidget()
+        header_layout = QHBoxLayout()
+        results_header.setLayout(header_layout)
+
+        results_label = QLabel()
+        header_layout.addWidget(results_label)
+
+        close_button = QPushButton()
+        close_button.setIcon(QIcon('resources/svg/spCloseIcon.svg'))
+        close_button.setToolTip('Close the search tab')
+        close_button.pressed.connect(self.remove_self)
+        header_layout.addStretch()
+        header_layout.addWidget(close_button)
+
+        results_widget_layout.addWidget(results_header)
+
+        # count the number of times the search term(s) was/were found in the record so that they can be sorted
+        filtered_results = []
+        for line in result_list:
+            counter = 0
+            for item in line[0]:
+                counter += 1
+
+            words_found = str(line[1])
+            words_found = words_found.replace('[', '')
+            words_found = words_found.replace(']', '')
+            words_found = words_found.replace("\'", '')
+            filtered_results.append((
+                str(line[0][0]),
+                str(line[2]),
+                words_found,
+                line[0][3],
+                line[0][16],
+                line[0][17],
+                line[0][21][0:100] + '...'))
+
+        model = QStandardItemModel(len(filtered_results), 5)
+        for i in range(len(filtered_results)):
+            for n in range(len(filtered_results[i])):
+                item = QStandardItem(filtered_results[i][n])
+                item.setEditable(False)
+                model.setItem(i, n, item)
+        model.setHeaderData(0, Qt.Orientation.Horizontal, 'ID')
+        model.setHeaderData(1, Qt.Orientation.Horizontal, '# of\r\nMatches')
+        model.setHeaderData(2, Qt.Orientation.Horizontal, 'Word(s) Found')
+        model.setHeaderData(3, Qt.Orientation.Horizontal, 'Sermon Text')
+        model.setHeaderData(4, Qt.Orientation.Horizontal, 'Sermon Title')
+        model.setHeaderData(5, Qt.Orientation.Horizontal, 'Sermon Date')
+        model.setHeaderData(6, Qt.Orientation.Horizontal, 'Sermon Snippet')
+
+        results_table_view = QTableView()
+        results_table_view.setModel(model)
+        results_table_view.setColumnWidth(0, 30)
+        results_table_view.setColumnWidth(1, 60)
+        results_table_view.setColumnWidth(2, 150)
+        results_table_view.setColumnWidth(3, 200)
+        results_table_view.setColumnWidth(4, 200)
+        results_table_view.setColumnWidth(5, 100)
+        results_table_view.setColumnWidth(6, 500)
+        results_table_view.setShowGrid(False)
+        results_table_view.setSelectionBehavior(QTableView.SelectionBehavior.SelectRows)
+        results_table_view.doubleClicked.connect(
+            lambda: self.retrieve_selection(model, results_table_view.selectionModel().currentIndex().row()))
+        results_widget_layout.addWidget(results_table_view)
+
+        if len(filtered_results) == 1:
+            results_label.setText(str(len(
+                filtered_results)) + ' result found.\nDouble-click a result below to open it.')
+        else:
+            results_label.setText(str(len(
+                filtered_results)) + ' results found.\nDouble-click a result below to open it.')
+
+    def retrieve_selection(self, model, selection):
+        """
+        Method to pull up whichever record the user selects
+
+        :param QStandardItemModel model: The model applied to the results_table_view
+        :param int selection: The row number of the results_table_view that was double-clicked
+        """
+        # be sure to check for changes before pulling up the new record
+        goon = True
+        if self.gui.changes:
+            goon = self.gui.main.ask_save()
+        if goon:
+            id_ = model.index(selection, 0).data()
+            index = 0
+            for item in self.gui.main.ids:
+                if int(item) == int(id_):
+                    break
+                index += 1
+            self.gui.main.get_by_index(index)
+            self.gui.tabbed_frame.setCurrentWidget(self.gui.tabbed_frame.widget(0))
+
+    def remove_self(self):
+        """
+        Method to remove this widget's tab from the GUI's tabbed widget.
+        """
+        self.gui.tabbed_frame.removeTab(5)
+        self.gui.tabbed_frame.setCurrentWidget(self.gui.tabbed_frame.widget(0))
+        self.destroy()
+
+
+class SermonView(QWidget):
+    def __init__(self, gui, text):
+        super().__init__()
+        self.gui = gui
+        self.text = text
+        self.button_bar = QWidget()
+        self.next_page_button = QPushButton('next')
+        self.previous_page_button = QPushButton('previous')
+        self.sermon_label = QLabel()
+        self.num_pages = 0
+        self.current_page = 0
+        self.current_font = QFont(
+            self.gui.main.user_settings['font_family'], int(self.gui.main.user_settings['font_size']))
+        self.page_label = None
+
+        self.init_components()
+        self.showMaximized()
+        self.gui.main.app.processEvents()
+
+        self.page_width = self.width() - 40
+        self.page_height = self.height() - self.button_bar.height() - 40
+
+        self.page_pixmaps = []
+        self.create_document_pages()
+        self.set_page()
+
+        self.previous_page_button.setEnabled(False)
+        if len(self.page_pixmaps) < 2:
+            self.next_page_button.setEnabled(False)
+        self.add_page_buttons()
+
+    def init_components(self):
+        self.setParent(self.gui)
+        self.setWindowFlag(Qt.WindowType.Window)
+        self.setWindowTitle('Sermon Viewer')
+
+        layout = QVBoxLayout(self)
+
+        button_layout = QHBoxLayout(self.button_bar)
+        layout.addWidget(self.button_bar)
+
+        zoom_out_button = QPushButton()
+        zoom_out_button.setObjectName('zoom_out')
+        zoom_out_button.setIcon(QIcon('resources/svg/spZoomOut.svg'))
+        zoom_out_button.setIconSize(QSize(36, 36))
+        zoom_out_button.pressed.connect(self.zoom)
+        button_layout.addWidget(zoom_out_button)
+        button_layout.addSpacing(20)
+
+        zoom_in_button = QPushButton()
+        zoom_in_button.setObjectName('zoom_in')
+        zoom_in_button.setIcon(QIcon('resources/svg/spZoomIn.svg'))
+        zoom_in_button.setIconSize(QSize(36, 36))
+        zoom_in_button.pressed.connect(self.zoom)
+        button_layout.addWidget(zoom_in_button)
+        button_layout.addStretch()
+
+        self.page_label = QLabel()
+        self.page_label.setFont(self.gui.bold_font)
+        button_layout.addWidget(self.page_label)
+        button_layout.addStretch()
+
+        layout.addWidget(self.sermon_label)
+        layout.addStretch()
+
+    def add_page_buttons(self):
+        page_button_next_stylesheet = (
+            'QPushButton {'
+            '   border: none;'
+            '   background: transparent;'
+            '   color: #00000000;'
+            '}'
+            'QPushButton:hover {'
+            '   background-color: rgba(255, 255, 255, 150);'
+            '   background-image: url("resources/svg/spNextPageIcon.svg");'
+            '   background-repeat: no-repeat;'
+            '   background-position: center;'
+            '}'
+            'QPushButton:pressed {'
+            '   background-color: #ffffff;'
+            '   color: #000000;'
+            '}'
+        )
+        page_button_previous_stylesheet = (
+            'QPushButton {'
+            '   border: none;'
+            '   background: transparent;'
+            '   color: #00000000;'
+            '}'
+            'QPushButton:hover {'
+            '   background-color: rgba(255, 255, 255, 150);'
+            '   background-image: url("resources/svg/spPrevPageIcon.svg");'
+            '   background-repeat: no-repeat;'
+            '   background-position: center;'
+            '}'
+            'QPushButton:pressed {'
+            '   background-color: #ffffff;'
+            '   color: #000000;'
+            '}'
+        )
+
+        self.previous_page_button.setParent(self)
+        self.previous_page_button.setObjectName('previous')
+        self.previous_page_button.setGeometry(
+            0, self.height() - self.page_height - 20, 200, self.page_height)
+        self.previous_page_button.setStyleSheet(page_button_previous_stylesheet)
+        self.previous_page_button.pressed.connect(self.set_page)
+        self.previous_page_button.show()
+
+        self.next_page_button.setParent(self)
+        self.next_page_button.setObjectName('next')
+        self.next_page_button.setGeometry(self.width() - 200, self.height() - self.page_height - 20, 200, self.page_height)
+        self.next_page_button.setStyleSheet(page_button_next_stylesheet)
+        self.next_page_button.pressed.connect(self.set_page)
+        self.next_page_button.show()
+
+    def create_document_pages(self):
+        document = QTextDocument()
+        document.setPageSize(QSizeF(self.page_width, self.page_height))
+        document.setDefaultFont(self.current_font)
+
+        text_option = QTextOption()
+        text_option.setWrapMode(QTextOption.WrapMode.WordWrap)
+        document.setDefaultTextOption(text_option)
+
+        document.setHtml(self.text)
+
+        self.page_pixmaps = []
+        current_y = 0
+        painter = QPainter()
+        self.num_pages = document.pageCount()
+        for i in range(document.pageCount()):
+            pixmap = QPixmap(self.page_width, self.page_height)
+            pixmap.fill(Qt.GlobalColor.white)
+
+            painter.begin(pixmap)
+            painter.translate(0, -current_y)
+            document.drawContents(painter)
+            painter.end()
+
+            self.page_pixmaps.append(pixmap)
+            current_y += self.page_height
+
+    def zoom(self):
+        if self.sender().objectName() == 'zoom_in':
+            self.current_font.setPointSize(self.current_font.pointSize() + 2)
+        else:
+            self.current_font.setPointSize(self.current_font.pointSize() - 2)
+        self.create_document_pages()
+        self.set_page()
+
+    def set_page(self):
+        if self.sender().objectName() == 'next':
+            self.current_page += 1
+        elif self.sender().objectName() == 'previous':
+            self.current_page -= 1
+
+        if self.current_page == 0:
+            self.previous_page_button.setEnabled(False)
+        else:
+            self.previous_page_button.setEnabled(True)
+
+        if self.current_page >= len(self.page_pixmaps) - 1:
+            self.current_page = len(self.page_pixmaps) - 1
+            self.next_page_button.setEnabled(False)
+        else:
+            self.next_page_button.setEnabled(True)
+
+        self.sermon_label.setPixmap(self.page_pixmaps[self.current_page])
+        self.page_label.setText(f'Page {self.current_page + 1} of {self.num_pages}')
